break ./artifacts into more source files
