+++ workers/generic-worker/artifacts/link.go
@@ -1,0 +1,28 @@
+package artifacts
+
+import "github.com/taskcluster/taskcluster/v44/clients/client-go/tcqueue"
+
+type LinkArtifact struct {
+	*BaseArtifact
+	Artifact    string
+	ContentType string
+}
+
+func (linkArtifact *LinkArtifact) ProcessResponse(response interface{}, logger Logger) error {
+	logger.Infof("Uploading link artifact %v to artifact %v with expiry %v", linkArtifact.Name, linkArtifact.Artifact, linkArtifact.Expires)
+	// nothing to do
+	return nil
+}
+
+func (linkArtifact *LinkArtifact) RequestObject() interface{} {
+	return &tcqueue.LinkArtifactRequest{
+		Expires:     linkArtifact.Expires,
+		StorageType: "link",
+		ContentType: linkArtifact.ContentType,
+		Artifact:    linkArtifact.Artifact,
+	}
+}
+
+func (linkArtifact *LinkArtifact) ResponseObject() interface{} {
+	return new(tcqueue.LinkArtifactResponse)
+}
