+++ workers/generic-worker/artifacts/redirect.go
@@ -1,0 +1,28 @@
+package artifacts
+
+import "github.com/taskcluster/taskcluster/v44/clients/client-go/tcqueue"
+
+type RedirectArtifact struct {
+	*BaseArtifact
+	URL         string
+	ContentType string
+}
+
+func (redirectArtifact *RedirectArtifact) ProcessResponse(response interface{}, logger Logger) error {
+	logger.Infof("Uploading redirect artifact %v to URL %v with mime type %q and expiry %v", redirectArtifact.Name, redirectArtifact.URL, redirectArtifact.ContentType, redirectArtifact.Expires)
+	// nothing to do
+	return nil
+}
+
+func (redirectArtifact *RedirectArtifact) RequestObject() interface{} {
+	return &tcqueue.RedirectArtifactRequest{
+		ContentType: redirectArtifact.ContentType,
+		Expires:     redirectArtifact.Expires,
+		StorageType: "reference",
+		URL:         redirectArtifact.URL,
+	}
+}
+
+func (redirectArtifact *RedirectArtifact) ResponseObject() interface{} {
+	return new(tcqueue.RedirectArtifactResponse)
+}
